003_azure_security/secure_azure_solutions/requirements.txt
python-dotenv
azure-keyvault-secrets
azure-identity